% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvemm.R
\name{DIF_GVEMM}
\alias{DIF_GVEMM}
\title{GVEMM Algorithms for DIF Detection in 2PL Models}
\usage{
DIF_GVEMM(
  Y,
  D,
  X,
  Method = "IWGVEMM",
  Lambda0 = seq(0.2, 0.7, by = 0.1),
  iter = 1000,
  eps = 0.001,
  c = 0.7,
  S = 10,
  M = 10,
  lr = 0.1
)
}
\arguments{
\item{Y}{An N by J binary matrix of item responses}

\item{D}{A J by G binary matrix of loading indicators}

\item{X}{An N dimensional vector of group indicators (integers from 1 to G)}

\item{Method}{Estimation algorithm, one of `'GVEMM'` and `'IWGVEMM'`}

\item{Lambda0}{A vector of `lambda0` values for L1 penalty (`lambda` is `sqrt(N) * lambda0`)}

\item{iter}{Maximum number of iterations}

\item{eps}{Termination criterion on numerical accuracy}

\item{c}{Constant for computing GIC}

\item{S}{Sample size for approximating the expected lower bound (`'IWGVEMM'` only)}

\item{M}{Sample size for approximating a tighter lower bound (`'IWGVEMM'` only)}

\item{lr}{Learning rate for Adam (`'IWGVEMM'` only)}
}
\value{
A list whose length is equal to `Lambda0`:\tabular{ll}{
\code{lambda0} \tab {Corresponding element in \code{Lambda0}} \cr
\tab \cr
\code{lambda} \tab {\code{sqrt(N) * lambda0}} \cr
\tab \cr
\code{iter} \tab {Number(s) of iterations} \cr
\tab \cr
\code{SIGMA} \tab {Person-level posterior covariance matrices} \cr
\tab \cr
\code{MU} \tab {Person-level posterior mean vectors} \cr
\tab \cr
\code{Sigma} \tab {Group-level posterior covariance matrices} \cr
\tab \cr
\code{Mu} \tab {Group-level posterior mean vectors} \cr
\tab \cr
\code{a} \tab {Slopes for group 1} \cr
\tab \cr
\code{b} \tab {Intercepts for group 1} \cr
\tab \cr
\code{gamma} \tab {DIF parameters for the slopes} \cr
\tab \cr
\code{beta} \tab {DIF parameters for the intercepts} \cr
\tab \cr
\code{ll} \tab {Log-likelihood} \cr
\tab \cr
\code{l0} \tab {Number of nonzero parameters in \code{gamma} and \code{beta}} \cr
\tab \cr
\code{AIC} \tab {Akaike Information Criterion} \cr
\tab \cr
\code{BIC} \tab {Bayesian Information Criterion} \cr
\tab \cr
\code{GIC} \tab {Generalized Information Criterion} \cr
}
}
\description{
GVEMM Algorithms for DIF Detection in 2PL Models
}
\examples{
\dontrun{
result <- with(DIF_simdata, DIF_GVEMM(Y, D, X))}
}
