% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm.R
\name{DIF_LRT}
\alias{DIF_LRT}
\title{Likelihood Ratio Test for DIF Detection in 2PL Models}
\usage{
DIF_LRT(Y, D, X, Unif = F)
}
\arguments{
\item{Y}{An N by J binary matrix of item responses}

\item{D}{A J by G binary matrix of loading indicators}

\item{X}{An N dimensional vector of group indicators (integers from 1 to G)}

\item{Unif}{Whether to detect uniform DIF only}
}
\value{
A list:\tabular{ll}{
\code{Sigma} \tab {Group-level posterior covariance matrices} \cr
\tab \cr
\code{Mu} \tab {Group-level posterior mean vectors} \cr
\tab \cr
\code{a} \tab {Slopes for group 1} \cr
\tab \cr
\code{b} \tab {Intercepts for group 1} \cr
\tab \cr
\code{gamma} \tab {DIF parameters for the slopes} \cr
\tab \cr
\code{beta} \tab {DIF parameters for the intercepts} \cr
}
}
\description{
Likelihood Ratio Test for DIF Detection in 2PL Models
}
\examples{
\dontrun{
result <- with(DIF_simdata, DIF_LRT(Y, D, X))}
}
