% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DIF_simdata}
\alias{DIF_simdata}
\title{Simulated Data Set}
\format{
A list of components of the data set:\tabular{ll}{
\code{Y} \tab {Item responses} \cr
\tab \cr
\code{D} \tab {Loading indicators} \cr
\tab \cr
\code{X} \tab {Group indicators} \cr
\tab \cr
\code{j} \tab {Number of DIF items (the first \code{j} items have DIF)} \cr
\tab \cr
\code{params} \tab {True parameters used for generating the data:} \cr
\tab \tabular{ll}{
\tab \cr
\code{a} \tab {Slopes} \cr
\tab \cr
\code{b} \tab {Intercepts} \cr
\tab \cr
\code{theta} \tab {Latent traits} \cr
}}
}
\usage{
DIF_simdata
}
\description{
Simulated Data Set
}
\keyword{datasets}
